+++ packages/components/CHANGELOG.md
@@ -19,6 +19,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 - `Popover`: support for custom slotted button
   ([#3861](https://github.com/porsche-design-system/porsche-design-system/pull/3861))
 
+#### Fixed
+
+- `Select Wrapper`: update of scroll position if dropdown is navigated with keyboard
+  ([#3858](https://github.com/porsche-design-system/porsche-design-system/pull/3858))
+
 ## [3.28.0-rc.3] - 2025-05-19
 
 ### Added